OvmfPkg: Add build support for lazy acceptance of memory
An SEV-SNP guest requires memory to be accepted before it can be used.
The idea behind lazy acceptance of memory is to improve the boot time
associated with an SEV-SNP guest and only accept memory below 4GB.

However, there are OS kernels that currently support SEV-SNP but don't
support un-accepted memory. Therefore, make lazy acceptance of memory a
build option.

Signed-off-by: Tom Lendacky <thomas.lendacky@amd.com>